+++ backend_springBoot/backend/src/main/java/app/project/entities/Projects.java
@@ -19,7 +19,7 @@ public class Projects{
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name="id")
-	private int ProjectId ;
+	private Long ProjectId ;
 	
 	@Column(name="name")
 	private String ProjectName  ;
